+++ src/main/java/com/mojang/minecraft/render/BlockData.java
@@ -18,9 +18,9 @@ public final class BlockData {
 	}
 	
 	public final float distanceSquared(Player var1) {
-		float var2 = var1.x - (float) this.x;
-		float var3 = var1.y - (float) this.y;
-		float var4 = var1.z - (float) this.z;
+		float var2 = var1.x - this.x;
+		float var3 = var1.y - this.y;
+		float var4 = var1.z - this.z;
 		return var2 * var2 + var3 * var3 + var4 * var4;
 	}
 }
